Copy readme.markdown from save-main branch
Originally this was supposed to be README.markdown so that it was still markdown rendered, but where .ma... was BEORE .md so it "overrode" the main landing document. I decided it's OK to have merge conflicts, as long as those merge conflicts are reasonably limited. For instance, don't rename the llvm directory, don't move everything to packages/*, and don't try to change the general structure. But, changing README.md, adding GitHub Actions, removing the PR thing, and other general customizations are OK. Basically just keep the primary code, change what you need, but don't needlessly customize.